pyproject.toml
[tool.poetry]
name = "livetalking"
version = "0.1.0"
description = "Real-time interactive streaming digital human"
authors = ["lipku"]
readme = "README.md"
packages = [{include = "livetalking_pkg"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
torch = ">=2.0.0"
torchvision = "*"
torchaudio = "*"
torch-ema = "*"
ninja = "*"
trimesh = "*"
opencv-python = "*"
tensorboardX = "*"
numpy = "*"
pandas = "*"
tqdm = "*"
matplotlib = "*"
PyMCubes = "*"
rich = "*"
dearpygui = "*"
packaging = "*"
scipy = "*"
scikit-learn = "*"
face_alignment = "*"
python_speech_features = "*"
numba = "*"
resampy = "*"
soundfile = "0.12.1"
sounddevice = "*"
einops = "*"
configargparse = "*"
lpips = "0.1.3"
imageio-ffmpeg = "*"
transformers = "4.46.2"
edge_tts = "*"
flask = "*"
flask_sockets = "*"
opencv-python-headless = "*"
aiortc = "*"
aiohttp_cors = "*"
ffmpeg-python = "*"
omegaconf = "*"
diffusers = "*"
accelerate = "*"
librosa = "*"
openai = "*"
websockets = "12.0"
gradio_client = "*"
azure-cognitiveservices-speech = "*"
mcp = "*"
python-dotenv = "^1.2.1"
httpx = "^0.28.1"
pydub = "^0.25.1"
requests = "^2.32.3"  # 用于豆包TTS HTTP调用
aiohttp = "^3.11.10"  # 异步HTTP客户端
websocket-client = "^1.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.11.0"
isort = "^5.12.0"
flake8 = "^6.1.0"

[tool.poetry.scripts]
livetalking = "livetalking_pkg.start_quick_fixed:main"
